Rename web-app-files components to follow essential naming conventions
