+++ packages/web-app-files/src/components/AppBar/SelectedResources/SizeInfo.vue
