+++ packages/web-app-files/src/components/SideBar/Actions/FileActions.vue
