+++ packages/web-app-files/src/components/SideBar/Links/CopyToClipboardButton.vue
