+++ packages/web-app-files/src/components/SideBar/Links/PublicLinks/LinkInfo.vue
