+++ packages/web-app-files/src/components/SideBar/Links/PublicLinks/ListItem.vue
@@ -9,14 +9,14 @@
 </template>
 
 <script>
-import LinkInfo from './Info.vue'
-import LinkActions from './Actions.vue'
+import LinkActions from './LinkActions.vue'
+import LinkInfo from './LinkInfo.vue'
 
 export default {
   name: 'PublicLinkListItem',
   components: {
-    LinkInfo,
-    LinkActions
+    LinkActions,
+    LinkInfo
   },
   props: {
     link: {
