+++ packages/web-app-files/src/components/SideBar/Shares/Collaborators/AdditionalPermissions.vue
