+++ packages/web-app-files/src/components/SideBar/Shares/Collaborators/NewCollaborator.vue
