+++ packages/web-app-files/src/components/Upload/DetailsWidget.vue
