+++ packages/web-app-files/tests/unit/components/Collaborators/AutocompleteItem.spec.js
@@ -1,6 +1,6 @@
 import { createLocalVue, shallowMount } from '@vue/test-utils'
 import Vuex from 'vuex'
-import AutocompleteItem from 'packages/web-app-files/src/components/Sidebar/Shares/Collaborators/AutocompleteItem.vue'
+import AutocompleteItem from 'packages/web-app-files/src/components/SideBar/Shares/Collaborators/AutocompleteItem.vue'
 import stubs from '../../../../../../tests/unit/stubs'
 
 const localVue = createLocalVue()
